requirements.txt
dash==3.0.3
Flask==3.0.3
flask-cors==5.0.1
numpy==2.2.5
pandas==2.2.3
plotly==6.0.1
python-dotenv==1.1.0
requests==2.32.3